requirements.txt
ciso8601==1.0.1
Flask==0.10.1
gunicorn==19.4.1
itsdangerous==0.24
Jinja2==2.8
jsonpickle==0.9.2
MarkupSafe==0.23
pip-autoremove==0.9.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylibmc==1.5.0
PyPubSub==3.3.0
pytz==2015.7
redis==2.10.5
requests==2.8.1
six==1.10.0
sleekxmpp==1.3.1
transitions==0.2.9
Werkzeug==0.11.2
wheel==0.24.0
